Add Discord data proxy for REST (#757)
* Connect rest to internal data proxy service

* Revert changes to postgresql dockerfile

* Remove build script in rest

* Remove spectacles/rest from tests, improve coverage

* Improve test coverage

* Expose port 4000 of api

+++ hasura/optional_migrations/1586479065565_next/up.sql
@@ -126,7 +126,7 @@ alter table guild_settings
 ;
 
 insert into guild_settings (
-	select guild as guild_id,
+	select guild::bigint as guild_id,
 		coalesce((settings ->> 'PREFIX')::text, '?') as prefix,
 		(settings ->> 'MOD_LOG')::bigint as mod_log_channel_id,
 		(settings ->> 'MOD_ROLE')::bigint as mod_role_id,
